requirements.txt
asgiref==3.8.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
django-crispy-forms==2.4
django-etc==1.4.0
django-jazzmin==3.0.1
django-ranged-response==0.2.0
django-secretballot==2.0.0
django-simple-captcha==0.6.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.94.2
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.4.2
flake8==7.3.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.0
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
